Revert "Bug 2032985 - Make PBrowser::RequestNativeKeyBindings more idiomatic. r=masayuki" for causing build bustages  EnumSerializer
This reverts commit fdbf57f92ff8591af0b0e02d96b3908601a7324d.

UltraBlame original commit: db0162e6e582803c7d1e48084450ae95ab311a01

+++ dom/ipc/BrowserChild.cpp
@@ -2396,7 +2396,8 @@ void BrowserChild::RequestEditCommands(NativeKeyBindingsType aType,
   
   
   WidgetKeyboardEvent localEvent(aEvent);
-  SendRequestNativeKeyBindings(aType, localEvent, &aCommands);
+  SendRequestNativeKeyBindings(static_cast<uint32_t>(aType), localEvent,
+                               &aCommands);
 }
 
 mozilla::ipc::IPCResult BrowserChild::RecvSynthesizedEventResponse(
